+++ SRC.Sharp/SRCCore/Units/Unit.feature.cs
@@ -12,20 +12,12 @@ namespace SRCCore.Units
     {
         public IList<Models.FeatureData> Features => colFeature.List;
 
-        //        // 特殊能力の総数
-        //        public int CountFeature()
-        //        {
-        //            int CountFeatureRet = default;
-        //            CountFeatureRet = colFeature.Count;
-        //            return CountFeatureRet;
-        //        }
-
         //        // 特殊能力
         //        public string Feature(string Index)
         //        {
         //            string FeatureRet = default;
         //            FeatureData fd;
-        //            fd = (FeatureData)colFeature[Index];
+        //            fd = (FeatureData);
         //            FeatureRet = fd.Name;
         //            return FeatureRet;
         //        }
@@ -33,191 +25,7 @@ namespace SRCCore.Units
         // 特殊能力の名称
         public string FeatureName(string Index)
         {
-            return FeatureNameInt(Index, colFeature);
-        }
-
-        private string FeatureNameInt(string Index, SrcCollection<FeatureData> feature_list)
-        {
-            FeatureData fd = feature_list[Index];
-            if (fd == null)
-            {
-                return Index;
-            }
-
-            // 非表示の能力
-            switch (fd.Name ?? "")
-            {
-                case "ノーマルモード":
-                case "パーツ合体":
-                case "換装":
-                case "制限時間":
-                case "制御不可":
-                case "主形態":
-                case "他形態":
-                case "合体制限":
-                case "格闘武器":
-                case "迎撃武器":
-                case "合体技":
-                case "変形技":
-                case "ランクアップ":
-                case "追加パイロット":
-                case "暴走時パイロット":
-                case "追加サポート":
-                case "装備個所":
-                case "ハードポイント":
-                case "武器クラス":
-                case "防具クラス":
-                case "ＢＧＭ":
-                case "武器ＢＧＭ":
-                case "アビリティＢＧＭ":
-                case "合体ＢＧＭ":
-                case "分離ＢＧＭ":
-                case "変形ＢＧＭ":
-                case "ハイパーモードＢＧＭ":
-                case "ユニット画像":
-                case "パイロット画像":
-                case "パイロット愛称":
-                case "パイロット読み仮名":
-                case "性別":
-                case "性格変更":
-                case "吸収":
-                case "無効化":
-                case "耐性":
-                case "弱点":
-                case "有効":
-                case "特殊効果無効化":
-                case "アイテム所有":
-                case "レアアイテム所有":
-                case "ラーニング可能技":
-                case "改造費修正":
-                case "最大改造数":
-                case "パイロット能力付加":
-                case "パイロット能力強化":
-                case "非表示":
-                case "攻撃属性":
-                case "射程延長":
-                case "武器強化":
-                case "命中率強化":
-                case "ＣＴ率強化":
-                case "特殊効果発動率強化":
-                case "必要技能":
-                case "不必要技能":
-                case "ダミーユニット":
-                case "地形ユニット":
-                case "召喚解除コマンド名":
-                case "変身解除コマンド名":
-                case "１人乗り可能":
-                case "特殊効果":
-                case "戦闘アニメ":
-                case "パイロット地形適応変更":
-                case "メッセージクラス":
-                case "用語名":
-                case "発光":
-                    // ユニット用特殊能力
-                    return "";
-
-                case "愛称変更":
-                case "読み仮名変更":
-                case "サイズ変更":
-                case "地形適応変更":
-                case "地形適応固定変更":
-                case "空中移動":
-                case "陸上移動":
-                case "水中移動":
-                case "宇宙移動":
-                case "地中移動":
-                case "修理費修正":
-                case "経験値修正":
-                case "最大弾数増加":
-                case "ＥＮ消費減少":
-                case "Ｖ－ＵＰ":
-                case "大型アイテム":
-                case "呪い":
-                    // アイテム用特殊能力
-                    return "";
-            }
-
-            // 拡大画像能力は「拡大画像(文字列)」といった指定もあるので他の非表示能力と異なる
-            // 判定方法を使う
-            if (Strings.InStr(fd.Name, "拡大画像") == 1)
-            {
-                return "";
-            }
-
-            string resultName;
-            if (Strings.Len(fd.StrData) > 0)
-            {
-                // 別名の指定あり
-                resultName = GeneralLib.ListIndex(fd.StrData, 1);
-                if (resultName == "非表示" | resultName == "解説")
-                {
-                    resultName = "";
-                }
-            }
-            else if (fd.Level == Constants.DEFAULT_LEVEL)
-            {
-                // レベル指定なし
-                resultName = fd.Name;
-            }
-            else if (fd.Level >= 0d)
-            {
-                // レベル指定あり
-                resultName = fd.Name + "Lv" + SrcFormatter.Format(fd.Level);
-                if (fd.Name == "射撃強化")
-                {
-                    if (CountPilot() > 0)
-                    {
-                        if (MainPilot().HasMana())
-                        {
-                            resultName = "魔力強化Lv" + SrcFormatter.Format(fd.Level);
-                        }
-                    }
-                }
-            }
-            else
-            {
-                // マイナスのレベル指定
-                switch (fd.Name ?? "")
-                {
-                    case "格闘強化":
-                        resultName = "格闘低下" + "Lv" + SrcFormatter.Format(Math.Abs(fd.Level));
-                        break;
-
-                    case "射撃強化":
-                        resultName = "射撃低下" + "Lv" + SrcFormatter.Format(Math.Abs(fd.Level));
-                        if (CountPilot() > 0)
-                        {
-                            if (MainPilot().HasMana())
-                            {
-                                resultName = "魔力低下Lv" + SrcFormatter.Format(Math.Abs(fd.Level));
-                            }
-                        }
-
-                        break;
-
-                    case "命中強化":
-                        resultName = "命中低下" + "Lv" + SrcFormatter.Format(Math.Abs(fd.Level));
-                        break;
-
-                    case "回避強化":
-                        resultName = "回避低下" + "Lv" + SrcFormatter.Format(Math.Abs(fd.Level));
-                        break;
-
-                    case "技量強化":
-                        resultName = "技量低下" + "Lv" + SrcFormatter.Format(Math.Abs(fd.Level));
-                        break;
-
-                    case "反応強化":
-                        resultName = "反応低下" + "Lv" + SrcFormatter.Format(Math.Abs(fd.Level));
-                        break;
-
-                    default:
-                        resultName = fd.Name + "Lv" + SrcFormatter.Format(fd.Level);
-                        break;
-                }
-            }
-
-            return resultName;
+            return colFeature[Index].FeatureName(this);
         }
 
         //        public string FeatureName0(string Index)
@@ -235,14 +43,13 @@ namespace SRCCore.Units
         // 特殊能力のレベル
         public double FeatureLevel(string Index)
         {
-            var level = colFeature[Index]?.Level ?? 0d;
-            return level == Constants.DEFAULT_LEVEL ? 1d : level;
+            return colFeature[Index]?.FeatureLevel ?? 0d;
         }
 
         // 特殊能力のデータ
         public string FeatureData(string Index)
         {
-            return colFeature[Index]?.StrData ?? "";
+            return colFeature[Index]?.Data ?? "";
         }
 
         //        // 特殊能力の必要技能
